[🔥Checkpoint] Context API
✅ Context Api Setup Todo
