+++ src/components/TodoApp/TodoItem.jsx
@@ -1,40 +1,28 @@
 /* eslint-disable */
 import { useState, useRef } from 'react';
+import { useTodosContext } from 'context/TodosContext';
 import styles from './TodoItem.module.css';
 import { FaTrash } from "react-icons/fa";
 import { AiFillEdit } from "react-icons/ai";
 
 
-const TodoItem = ({ itemProp, setTodos, delTodo, setUpdate }) => {
-  const [updateInput, setUpdateInput] = useState(itemProp.title);
-  //const editInputRef = useRef(null);
+
+const TodoItem = ({ itemProp }) => {
   const [editing, setEditing] = useState(false);
-  const handleChange = (id) => {
-    setTodos((prevState) => prevState.map((todo) => {
-      if (todo.id === id) {
-        return {
-          ...todo,
-          completed: !todo.completed,
-        };
-      }
-      return todo;
-    }));
-  };
-  const handleEditing = () => {
-    setEditing(true);
-  };
-  const handleUpdatedDone = (event) => {
-    if (event.key === 'Enter') {
-      // setUpdate(editInputRef.current.value, itemProp.id);
-      setUpdate(updateInput, itemProp.id);
-      setEditing(false);
-    }
-  };
+
+  const { handleChange, delTodo, setUpdate } = useTodosContext();
+
+  const editInputRef = useRef(null);
+
   const completedStyle = {
     fontStyle: 'italic',
     color: '#595959',
     opacity: 0.4,
     textDecoration: 'line-through',
+  };
+
+  const handleEditing = () => {
+    setEditing(true);
   };
 
   let viewMode = {};
@@ -45,6 +33,12 @@ const TodoItem = ({ itemProp, setTodos, delTodo, setUpdate }) => {
     editMode.display = 'none';
   }
 
+  const handleUpdatedDone = (event) => {
+    if (event.key === 'Enter') {
+      setUpdate(editInputRef.current.value, itemProp.id);
+      setEditing(false);
+    }
+  };
   return (
     <li className={styles.item}>
       <div className={styles.content} style={viewMode}>
@@ -53,22 +47,22 @@ const TodoItem = ({ itemProp, setTodos, delTodo, setUpdate }) => {
           checked={itemProp.completed}
           onChange={() => handleChange(itemProp.id)}
         />
-        <button type="button" onClick={handleEditing}>
-          <AiFillEdit style={{ color: "#5e5e5e", fontSize: "16px" }}/>
+        <button onClick={handleEditing}>
+          <AiFillEdit
+            style={{ color: '#5e5e5e', fontSize: '16px' }}
+          />
         </button>
-        <button type="button" onClick={() => delTodo(itemProp.id)}>
-          <FaTrash style={{ color: "#5e5e5e", fontSize: "16px" }}/>
+        <button onClick={() => delTodo(itemProp.id)}>
+          <FaTrash style={{ color: '#5e5e5e', fontSize: '16px' }} />
         </button>
         <span style={itemProp.completed ? completedStyle : null}>
-          {updateInput}
+          {itemProp.title}
         </span>
       </div>
       <input
         type="text"
-        value={updateInput}
-        onChange={(e) => setUpdateInput(e.target.value)}
-        // ref={editInputRef}
-        // defaultValue={itemProp.title}
+        ref={editInputRef}
+        defaultValue={itemProp.title}
         className={styles.textInput}
         style={editMode}
         onKeyDown={handleUpdatedDone}
