+++ src/components/TodoApp/TodosLogic.jsx
@@ -1,72 +1,12 @@
 import { TodosProvider } from 'context/TodosContext';
-import React, { useState, useEffect } from 'react';
-import { v4 as uuidv4 } from 'uuid';
 import InputTodo from './InputTodo';
 import TodoList from './TodoList';
 
-const TodosLogic = () => {
-  const [todos, setTodos] = useState(getInitialTodos());
-  useEffect(() => {
-    // Localstorage
-    const temp = JSON.stringify(todos);
-    localStorage.setItem('todos', temp);
-  }, [todos]);
-
-  // Get From Local Storage
-  function getInitialTodos() {
-    const temp = localStorage.getItem('todos');
-    const savedTodos = JSON.parse(temp);
-    return savedTodos || [];
-  }
-
-  const handleChange = (id) => {
-    setTodos((prevState) => prevState.map((todo) => {
-      if (todo.id === id) {
-        return {
-          ...todo,
-          completed: !todo.completed,
-        };
-      }
-      return todo;
-    }));
-  };
-  const delTodo = (id) => {
-    setTodos([
-      ...todos.filter((todo) => todo.id !== id),
-    ]);
-  };
-  const addTodoItem = (title) => {
-    const newTodo = {
-      id: uuidv4(),
-      title,
-      completed: false,
-    };
-    setTodos([...todos, newTodo]);
-  };
-
-  const setUpdate = (updatedTitle, id) => {
-    setTodos(
-      todos.map((todo) => {
-        if (todo.id === id) {
-          todo.title = updatedTitle;
-        }
-        return todo;
-      }),
-    );
-  };
-
-  return (
-    <TodosProvider>
-      <InputTodo addTodoItem={addTodoItem} />
-      <TodoList
-        todosProps={todos}
-        setTodos={setTodos}
-        handleChange={handleChange}
-        delTodo={delTodo}
-        setUpdate={setUpdate}
-      />
-    </TodosProvider>
-  );
-};
+const TodosLogic = () => (
+  <TodosProvider>
+    <InputTodo />
+    <TodoList />
+  </TodosProvider>
+);
 
 export default TodosLogic;
